deleted 0x09 not working:)

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -9,7 +9,7 @@
 
 int main(void)
 {
-	int i, f;
+	long int i, f;
 	f = 1;
 	
 	for (i = 0; i < 50; i++)
